+++ CHANGELOG.md
@@ -17,10 +17,13 @@ All notable changes to this project will be documented in this file.
          - New Usage Over Time chart - will show incrementing total filemant usage by day. 
          - Moved Re-Sync Filament Manager button to show with others (manage profile/spools).
      - New History Chart - Will show daily totals of success / cancelled and failed prints.
+     - Dashboard has the 3 new graphs detailed above available. 
       
 ### Changed
      - Fixed #228: Quick connect button will only activate once you have setup preferred connection preferences. 
      - Printer Logs Modal for OctoPrint logs now includes line number 
+     - Re-organised the System dashboard settings page to include new graph options
+     - History statistics now only take into account successful prints.
 
 ### Fixed
      - Fixed #226: Couldn't update OctoPrint instances, restart would get in the way.